Evict txs from the mempool if they haven't been committed after N blocks
Previously a tx was only removed from the mempool when the block it was
included in was committed by the node, or when the tx failed CheckTx
after a block was committed (which can only happen when recheck is
enabled).

Sometimes a previously committed tx is received from a peer via the
mempool reactor, possibly because the peer is lagging behind.
If recheck is disabled and the tx is not in the mempool cache then the
tx may end up sitting in the node's mempool forever.

To prevent txs from getting stuck in the mempool forever it's now
possible to specify a tx life window, this is the maximum number of
blocks the tx will be allowed to remain in the mempool. After the life
window expires the tx will be removed from the mempool, but will remain
in the mempool cache to prevent it from re-entering the mempool again.